Uses refactored code for part two
